@@ -1,6 +1,16 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import Home from '@/pages/home'
 import '@/styles/index.less'
 
-ReactDOM.render(<Home />, document.getElementById('root'))
+function Hello() {
+  return (
+    <div>
+      Hello world!
+      <Home />
+    </div>
+  )
+}
+
+const root = createRoot(document.getElementById('root')!)
+root.render(<Hello />)
