@@ -4,4 +4,4 @@ declare module '*.jpg'
 declare module '*.jpeg'
 declare module '*.css'
 declare module '*.less'
-declare module '*.syle'
+declare module '*.style'
